// lib/type.ts
export type color =
  | "default"
  | "defaultGray"
  | "primary"
  | "secondary"
  | "success"
  | "successDefault"
  | "info"
  | "warning"
  | "destructive";
export type InputColor =
  | "default"
  | "primary"
  | "secondary"
  | "success"
  | "info"
  | "warning"
  | "destructive"
  | "waringTwo";
export type shadow = "sm" | "md" | "lg" | "xl";
export type size = "default" | "sm" | "md" | "lg";
export type rounded = "sm" | "md" | "lg" | "full" | "none";
export type radius = "sm" | "md" | "lg" | "xl" | "none";

// config
export type layoutType = "vertical" | "horizontal" | "semi-box" | "compact";
export type contentType = "wide" | "boxed";
export type skinType = "default" | "bordered";
export type sidebarType = "classic" | "draggable" | "two-column" | "compact";
export type navBarType = "floating" | "sticky" | "hidden" | "default";
export type headerColorType = "default" | "coloured" | "transparent";